requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
seaborn
scikit-learn
pandas
opencv-python